+++ src/main/java/com/zdzc/electrocar/dto/GPSDto.java
@@ -36,7 +36,7 @@ public class GPSDto implements Serializable{
     /**
      * Acc状态
      */
-    private byte accStatus;
+    private int accStatus;
 
     public String getDeviceId() {
         return deviceId;
@@ -94,11 +94,11 @@ public class GPSDto implements Serializable{
         this.speed = speed;
     }
 
-    public byte getAccStatus() {
+    public int getAccStatus() {
         return accStatus;
     }
 
-    public void setAccStatus(byte accStatus) {
+    public void setAccStatus(int accStatus) {
         this.accStatus = accStatus;
     }
 }
